requirements.txt
streamlit
requests
beautifulsoup4
pandas
numpy
azure-ai-ml
azure-identity
python-dotenv